match: Format entire ODP port number in match_format().
struct flow has a union for the in_port field, and the two different
members of the union have different sizes: ofp_port_t is 16 bits,
odp_port_t is 32 bits.  On little-endian machines this doesn't matter much,
but on big-endian machines it's important to distinguish between them
because a small number in odp_port_t will show up as zero when viewed
through ofp_port_t.

There are in fact few parts of OVS that can handle a struct flow that
has either formats; most parts of OVS work with one or the other but not
both.  match_format(), however, does need to handle both cases, and it did
not: it displayed small odp_port_t values as zero on big-endian platforms.
This commit fixes the problem by checking the mask, which indicates which
format is in use, and displaying the appropriate field from the union.

Reported-by: Dimitri John Ledkov <xnox@ubuntu.com>
Reported-at: http://openvswitch.org/pipermail/discuss/2016-January/020072.html
Signed-off-by: Ben Pfaff <blp@ovn.org>